+++ CHANGELOG.md
@@ -7,6 +7,42 @@ and this project adheres to [CalVer](https://calver.org/).
 ## [UNRELEASED]
 ### Added
 - The awesome Leaflet plugin Leaflet.markercluster is added, so by setting the meta property `use_clustering` to `true` clustering can be enabled on single point vector layers. No other setting for Leaflet.markercluster is available for now.
+- With the `cssFiles` config it's possible to load external css file from the `configUrl`. E.g.:
+```json
+{
+  "cssFiles": [
+       "myStyles1.css",
+       "myStyles2.css"
+  ]
+}
+```
+- Conflict is now using `sqlQuery.js` to show info-click result. So now are multiple results possible. The `sqlQuery.js` module now has a Simple mode, a prefix for DOM elements to render in and callback for when selecting a row.
+- New config setting `dontUseAdvancedBaseLayerSwitcher` which will disable the 'Display two layers at once' option in the base layer switcher.
+- New GC2 Meta properties `select_function` which adds an onSelect callback to the result list in click-for-info.
+- It now possible to render an image carousel in popups from a click-for-info. Create a JSON field with a value like this and set Content to Image:
+```json
+[
+  {
+    "src": "https://image1.jpeg",
+    "att": "It's an image!"
+  },
+  {
+    "src": "https://image1.jpeg",
+    "att": "It's another image!!"
+  }
+]
+```
+- URL's can now be ignored for caching in Service Worker by using the `urlsIgnoredForCaching` setting in `config/config.js`. This setting can only be set in build time. E.g.:
+```json
+{
+    "urlsIgnoredForCaching": [
+        {
+            "regExp": true,
+            "requested": "part_of_the_url"
+        }
+    ]
+}
+```
 - Print setup can now be "sticky" (by using the state module), so it will not reset when re-activating the module. Stickiness can be toggled and off. Off is default.
 - Print setup is now stored in state snapshots. After state snapshot is activated the print setup will use the stored settings. The sticky toggle must be set to on or else the default print settings will be used.
 - New print API `/api/print/[database]/?state=[state id]` which will return the stored print from a snapshot as PNG (PDF is coming). The print will be created on the fly.
@@ -22,11 +58,13 @@ and this project adheres to [CalVer](https://calver.org/).
     - Layers with only one report column will be printed as a `|` separated string and not a table with one column.
     - New button for setting the print extent before creating a PDF.
     - Mouse click when releasing a rectangle/circle drag is suppressed.
-- Indicator in layer tree showing if a tile layer (MapServer or QGIS Server) is visible in the view extent. The Leaflet layer canvas element is checked for colored pixel. Be aware of the canvas being bigger than the view extent because of the buffer. A event is triggered when a layer changes visibility called `tileLayerVisibility:layers` with a payload like this:
-```javascript
+    - Multiple results from a Select-Object-To-Search-With is now possible. The standard `sqlQuery` is used to create list. 
+    
+- Indicator in layer tree showing if a tile layer (MapServer or QGIS Server) is visible in the view extent. The Leaflet layer canvas element is being checked for colored pixels. Be aware of the canvas being bigger than the view extent because of the buffer. A event is triggered when a layer changes visibility called `tileLayerVisibility:layers` with a payload like this:
+```json
 {
-    id: "schema.layer.geom",
-    dataIsVisible: true
+    "id": "schema.layer.geom",
+    "dataIsVisible": true
 }
 ```
 
@@ -43,7 +81,7 @@ and this project adheres to [CalVer](https://calver.org/).
 - CalVer is now used with month identifier like this: YYYY.MM.Minor.Modifier
 - Custom searches can now be added to danish search module.
 - `embed.js` will wait with loading Vidi until target element is visible in the DOM. This way, Vidi can be embedded in a element with `display:none`.
-- Its now possible to add custom extra searches to `danish.js`. A search needs a Elasticsearch index, which must have a id and string property. The latter is the search string. Also a look-up table/view with geometries is required. An example of a setup:
+- Its now possible to add custom extra searches to `danish.js`. A search needs an Elasticsearch index, which must have an id and string property. The latter is the search string. Also a look-up table/view with geometries is required. An example of a setup:
 ```javascript
 {
     searchConfig: {